patched nav-bar to allow scrolling content
angular-material provide a component for navigation closely
related to tab-bars, but with support for url locations.
The problem arise when the content overflows, as tab-bars
paginate nicely, but nav-bars just clamp the content.

The angular-material team said they will not support the
pagination feature in nav-bar, and the only avalible solution
is to monkey patch it to scroll inside a wrapper.

See the feature request:
https://github.com/angular/material/issues/8701